NGA-389: creating new python bindings to server api

+++ py-rest-client/product_catalogue_py_rest_client/models/product_l0_src.py
@@ -34,7 +34,7 @@ class ProductL0Src(object):
                             and the value is json key in definition.
     """
     openapi_types = {
-        'id': 'float',
+        'id': 'int',
         'l0_instrument_location': 'str',
         'metadata_persistent_id': 'str',
         'name': 'str',
@@ -78,7 +78,7 @@ class ProductL0Src(object):
 
 
         :return: The id of this ProductL0Src.  # noqa: E501
-        :rtype: float
+        :rtype: int
         """
         return self._id
 
@@ -88,7 +88,7 @@ class ProductL0Src(object):
 
 
         :param id: The id of this ProductL0Src.  # noqa: E501
-        :type: float
+        :type: int
         """
         if self.local_vars_configuration.client_side_validation and id is None:  # noqa: E501
             raise ValueError("Invalid value for `id`, must not be `None`")  # noqa: E501
