+++ py-rest-client/product_catalogue_py_rest_client/models/survey_l0_relation.py
@@ -34,7 +34,7 @@ class SurveyL0Relation(object):
                             and the value is json key in definition.
     """
     openapi_types = {
-        'id': 'float',
+        'id': 'int',
         'product_l0_src': 'ProductL0Src',
         'survey': 'Survey'
     }
@@ -66,7 +66,7 @@ class SurveyL0Relation(object):
 
 
         :return: The id of this SurveyL0Relation.  # noqa: E501
-        :rtype: float
+        :rtype: int
         """
         return self._id
 
@@ -76,7 +76,7 @@ class SurveyL0Relation(object):
 
 
         :param id: The id of this SurveyL0Relation.  # noqa: E501
-        :type: float
+        :type: int
         """
         if self.local_vars_configuration.client_side_validation and id is None:  # noqa: E501
             raise ValueError("Invalid value for `id`, must not be `None`")  # noqa: E501
